Move utils/settings to comms
